+++ src/main/java/ru/hh/resumebuilderbot/question/storage/graph/node/QuestionNodeTerminal.java
@@ -1,30 +1,28 @@
-package ru.hh.resumebuilderbot.question.storage.node.basic;
+package ru.hh.resumebuilderbot.question.storage.graph.node;
 
 import ru.hh.resumebuilderbot.Answer;
 import ru.hh.resumebuilderbot.question.Question;
-import ru.hh.resumebuilderbot.question.storage.node.QuestionNode;
+import ru.hh.resumebuilderbot.texts.storage.TextId;
+import ru.hh.resumebuilderbot.texts.storage.TextsStorage;
 import ru.hh.resumebuilderbot.user.data.storage.UserData;
 
 import java.util.Map;
 
-public class QuestionNodeLinear implements QuestionNode {
+public class QuestionNodeTerminal implements QuestionNode {
     private Question question;
-    private boolean isSkippable;
-    private QuestionNode next;
 
-    public QuestionNodeLinear(Question question, boolean isSkippable) {
-        this.question = question;
-        this.isSkippable = isSkippable;
+    public QuestionNodeTerminal() {
+        question = new Question(TextsStorage.getText(TextId.FINISHED));
     }
 
     @Override
     public void setLinks(Map<String, QuestionNode> links) {
-        next = links.get("next");
+
     }
 
     @Override
     public boolean answerIsValid(Answer answer) {
-        return question.answerIsAllowed(answer);
+        return true;
     }
 
     @Override
@@ -39,17 +37,17 @@ public class QuestionNodeLinear implements QuestionNode {
 
     @Override
     public QuestionNode getNext() {
-        return next;
+        return this;
     }
 
     @Override
     public boolean isSkippable() {
-        return isSkippable;
+        return true;
     }
 
     @Override
     public QuestionNode cloneContent() {
-        return new QuestionNodeLinear(question, isSkippable);
+        return new QuestionNodeTerminal();
     }
 
     @Override
