+++ src/main/java/ru/hh/resumebuilderbot/question/storage/graph/xml/parser/XMLParser.java
@@ -1,4 +1,4 @@
-package ru.hh.resumebuilderbot.question.storage.builder.xml.parser;
+package ru.hh.resumebuilderbot.question.storage.graph.xml.parser;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
