+++ src/main/java/ru/hh/resumebuilderbot/user/data/storage/UserData.java
@@ -3,8 +3,8 @@ package ru.hh.resumebuilderbot.user.data.storage;
 import ru.hh.resumebuilderbot.Answer;
 import ru.hh.resumebuilderbot.question.Question;
 import ru.hh.resumebuilderbot.question.storage.QuestionsStorage;
-import ru.hh.resumebuilderbot.question.storage.builder.Graph;
-import ru.hh.resumebuilderbot.question.storage.node.QuestionNode;
+import ru.hh.resumebuilderbot.question.storage.graph.Graph;
+import ru.hh.resumebuilderbot.question.storage.graph.node.QuestionNode;
 
 import java.util.ArrayList;
 import java.util.List;
